+++ Screens/Recorder/Workout/WorkoutHome.js
@@ -10,7 +10,7 @@ import { Text, View, StyleSheet, Image, BackHandler, Platform} from 'react-nativ
 import { Button, SideMenu, List, ListItem, Icon, Header, Divider } from 'react-native-elements'
 
 
-export class WorkHomeScreen extends React.Component{
+export class WorkoutHomeScreen extends React.Component{
   componentWillMount(){
     console.log("Current Screen: " + this.props.navigation.state.routeName)
   }